+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log
 
+## 2.1.5
+
+### Patch Changes
+
+🐛 Update golang-jwt/jwt dependency to v4.5.2
+🐛 Get default http transport from plugin-sdk-go
+
 ## [2.1.4]
 
 - **Fix** - Workflow runs - date filter now filters by time
+- **Fix** - Panic in project query
+- **Security** - Bump prismjs to 1.30.0
 
 ## [2.1.3]